requirements.txt
pre-commit
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
